+++ CHANGELOG.md
@@ -11,6 +11,10 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 - Hidden Paths property.
 
+### Fixed
+
+- Skeleton display.
+
 ## [2.21.0] - 2026-07-02
 
 ### Added